+++ website/landing/components/common/List.tsx
@@ -1,0 +1,7 @@
+import { styled } from "../../stitches.config";
+
+export const List = styled("ul", {
+  listStyle: "none",
+  margin: 0,
+  padding: 0,
+});