+++ website/docs/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.7](https://github.com/codesandbox/sandpack/compare/v0.3.6...v0.3.7) (2021-11-16)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.3.6](https://github.com/codesandbox/sandpack/compare/v0.3.5...v0.3.6) (2021-11-16)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.3.5](https://github.com/codesandbox/sandpack/compare/v0.3.4...v0.3.5) (2021-11-15)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.3.3](https://github.com/codesandbox/sandpack/compare/v0.3.2...v0.3.3) (2021-11-15)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.3.2](https://github.com/codesandbox/sandpack/compare/v0.3.1...v0.3.2) (2021-11-15)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+## [0.3.1](https://github.com/codesandbox/sandpack/compare/v0.3.0...v0.3.1) (2021-11-15)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
+# [0.3.0](https://github.com/codesandbox/sandpack/compare/v0.2.3...v0.3.0) (2021-11-15)
+
+**Note:** Version bump only for package sandpack-docs
+
+
+
+
+
 ## [0.2.3](https://github.com/codesandbox/sandpack/compare/v0.2.2...v0.2.3) (2021-11-11)
 
 **Note:** Version bump only for package sandpack-docs